2.Farm/6-keep-your-plant-secure/device_requirements.txt
azure-iot-device
counterfit
counterfit-shims-grove